Setup ruby on rails framework
I've chosen ruby on rails because I'm familiar with it.

I have also considered the following options:
- sinatra
- hanami

I did not want to install the whole ruby on rails ecosystem, so I have
started the project with the following command:

```
 rails new pistachio --minimal -T
```

This removes most of the libraries, e.g: mailer
The -T option skips installing tests, I want to use RSpec

I have chosen to keep SQlite as the default database since this is
a programming exercise and making someone else install PostgreSQL can
be a lengthy process.

The Ruby on Rails framework enables anyone to change the database,
a few things to keep in mind is to update the following files:

- config/database.yml
- Gemfile (replace sqlite3 gem with pg gem)

I have added a config/database_postgresql.yml, if you are interested
then just delete config/database.yml and rename accordingly.